# api_service/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  handler.AdminUserResponse:
    properties:
      api_key:
        type: string
      created_at:
        type: string
      id:
        type: integer
      last_synced_at:
        type: string
      remark:
        type: string
      sharing_enabled:
        type: boolean
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API server for the Cookie Syncer browser extension.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Cookie Syncer API
  version: "1.0"
paths:
  /admin/users:
    post:
      consumes:
      - application/json
      description: Creates new users based on the request body array. Each object
        in the array can specify a remark.
      parameters:
      - description: Array of users to create. Can be empty to create one default
          user.
        in: body
        name: body
        required: true
        schema:
          items:
            properties:
              remark:
                type: string
            type: object
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Returns an array of created users including their new API keys.
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handler.AdminUserResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - AdminKeyAuth: []
      summary: '[Admin] Create one or more users'
      tags:
      - Admin
  /admin/users/{id}:
    put:
      consumes:
      - application/json
      description: Updates a user's details, such as their remark.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User details to update
        in: body
        name: body
        required: true
        schema:
          properties:
            remark:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - AdminKeyAuth: []
      summary: '[Admin] Update user by ID'
      tags:
      - Admin
  /admin/users/{id}/refresh-key:
    post:
      description: Generates a new API key for the specified user and returns the
        full updated user object.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.AdminUserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - AdminKeyAuth: []
      summary: '[Admin] Refresh user API key by ID'
      tags:
      - Admin
  /admin/users/by-key/{apiKey}:
    put:
      consumes:
      - application/json
      description: Updates a user's details, such as their remark, by finding them
        via API key.
      parameters:
      - description: User API Key
        in: path
        name: apiKey
        required: true
        type: string
      - description: User details to update
        in: body
        name: body
        required: true
        schema:
          properties:
            remark:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - AdminKeyAuth: []
      summary: '[Admin] Update user by API key'
      tags:
      - Admin
  /admin/users/by-key/{apiKey}/refresh-key:
    post:
      description: Generates a new API key for the specified user and returns the
        full updated user object.
      parameters:
      - description: User API Key
        in: path
        name: apiKey
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/handler.AdminUserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - AdminKeyAuth: []
      summary: '[Admin] Refresh user API key by API key'
      tags:
      - Admin
...
securityDefinitions:
  AdminKeyAuth:
    in: header
    name: x-admin-key
    type: apiKey
  ApiKeyAuth:
    in: header
    name: x-api-key
    type: apiKey
  PoolKeyAuth:
    in: header
    name: x-pool-key
    type: apiKey
swagger: "2.0"